% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontal2D.R
\docType{data}
\name{frontal2D}
\alias{frontal2D}
\title{Frontal lobe functional connectivity in ADHD}
\format{A data frame with 48 rows and 381 variables:
\describe{
  \item{Group}{Diagnostic group, control or patient.}
  \item{Sex}{Sex factor, female (F) of male (M).}
  \item{Age}{Chronological age in years old after birth, numeric.}
  \item{FAG.FAD}{Functional connectivity between FAG and FAD regions, numeric}
  \item{FAG.F1G}{Functional connectivity between FAG and F1G regions, numeric}
  \item{FAD.F1G}{Functional connectivity between FAD and F1G regions, numeric}
  \item{FAG.F1D}{Functional connectivity between FAG and F1D regions, numeric}
  \item{FAD.F1D}{Functional connectivity between FAD and F1D regions, numeric}
  \item{F1G.F1D}{Functional connectivity between F1G and F1D regions, numeric}
  \item{FAG.F1OG}{Functional connectivity between FAG and F1OG regions, numeric}
  \item{FAD.F1OG}{Functional connectivity between FAD and F1OG regions, numeric}
  \item{F1G.F1OG}{Functional connectivity between F1G and F1OG regions, numeric}
  \item{F1D.F1OG}{Functional connectivity between F1D and F1OG regions, numeric}
  \item{FAG.F1OD}{Functional connectivity between FAG and F1OD regions, numeric}
  \item{FAD.F1OD}{Functional connectivity between FAD and F1OD regions, numeric}
  \item{F1G.F1OD}{Functional connectivity between F1G and F1OD regions, numeric}
  \item{F1D.F1OD}{Functional connectivity between F1D and F1OD regions, numeric}
  \item{F1OG.F1OD}{Functional connectivity between F1OG and F1OD regions, numeric}
  \item{FAG.F2G}{Functional connectivity between FAG and F2G regions, numeric}
  \item{FAD.F2G}{Functional connectivity between FAD and F2G regions, numeric}
  \item{F1G.F2G}{Functional connectivity between F1G and F2G regions, numeric}
  \item{F1D.F2G}{Functional connectivity between F1D and F2G regions, numeric}
  \item{F1OG.F2G}{Functional connectivity between F1OG and F2G regions, numeric}
  \item{F1OD.F2G}{Functional connectivity between F1OD and F2G regions, numeric}
  \item{FAG.F2D}{Functional connectivity between FAG and F2D regions, numeric}
  \item{FAD.F2D}{Functional connectivity between FAD and F2D regions, numeric}
  \item{F1G.F2D}{Functional connectivity between F1G and F2D regions, numeric}
  \item{F1D.F2D}{Functional connectivity between F1D and F2D regions, numeric}
  \item{F1OG.F2D}{Functional connectivity between F1OG and F2D regions, numeric}
  \item{F1OD.F2D}{Functional connectivity between F1OD and F2D regions, numeric}
  \item{F2G.F2D}{Functional connectivity between F2G and F2D regions, numeric}
  \item{FAG.F2OG}{Functional connectivity between FAG and F2OG regions, numeric}
  \item{FAD.F2OG}{Functional connectivity between FAD and F2OG regions, numeric}
  \item{F1G.F2OG}{Functional connectivity between F1G and F2OG regions, numeric}
  \item{F1D.F2OG}{Functional connectivity between F1D and F2OG regions, numeric}
  \item{F1OG.F2OG}{Functional connectivity between F1OG and F2OG regions, numeric}
  \item{F1OD.F2OG}{Functional connectivity between F1OD and F2OG regions, numeric}
  \item{F2G.F2OG}{Functional connectivity between F2G and F2OG regions, numeric}
  \item{F2D.F2OG}{Functional connectivity between F2D and F2OG regions, numeric}
  \item{FAG.F2OD}{Functional connectivity between FAG and F2OD regions, numeric}
  \item{FAD.F2OD}{Functional connectivity between FAD and F2OD regions, numeric}
  \item{F1G.F2OD}{Functional connectivity between F1G and F2OD regions, numeric}
  \item{F1D.F2OD}{Functional connectivity between F1D and F2OD regions, numeric}
  \item{F1OG.F2OD}{Functional connectivity between F1OG and F2OD regions, numeric}
  \item{F1OD.F2OD}{Functional connectivity between F1OD and F2OD regions, numeric}
  \item{F2G.F2OD}{Functional connectivity between F2G and F2OD regions, numeric}
  \item{F2D.F2OD}{Functional connectivity between F2D and F2OD regions, numeric}
  \item{F2OG.F2OD}{Functional connectivity between F2OG and F2OD regions, numeric}
  \item{FAG.F3OPG}{Functional connectivity between FAG and F3OPG regions, numeric}
  \item{FAD.F3OPG}{Functional connectivity between FAD and F3OPG regions, numeric}
  \item{F1G.F3OPG}{Functional connectivity between F1G and F3OPG regions, numeric}
  \item{F1D.F3OPG}{Functional connectivity between F1D and F3OPG regions, numeric}
  \item{F1OG.F3OPG}{Functional connectivity between F1OG and F3OPG regions, numeric}
  \item{F1OD.F3OPG}{Functional connectivity between F1OD and F3OPG regions, numeric}
  \item{F2G.F3OPG}{Functional connectivity between F2G and F3OPG regions, numeric}
  \item{F2D.F3OPG}{Functional connectivity between F2D and F3OPG regions, numeric}
  \item{F2OG.F3OPG}{Functional connectivity between F2OG and F3OPG regions, numeric}
  \item{F2OD.F3OPG}{Functional connectivity between F2OD and F3OPG regions, numeric}
  \item{FAG.F3OPD}{Functional connectivity between FAG and F3OPD regions, numeric}
  \item{FAD.F3OPD}{Functional connectivity between FAD and F3OPD regions, numeric}
  \item{F1G.F3OPD}{Functional connectivity between F1G and F3OPD regions, numeric}
  \item{F1D.F3OPD}{Functional connectivity between F1D and F3OPD regions, numeric}
  \item{F1OG.F3OPD}{Functional connectivity between F1OG and F3OPD regions, numeric}
  \item{F1OD.F3OPD}{Functional connectivity between F1OD and F3OPD regions, numeric}
  \item{F2G.F3OPD}{Functional connectivity between F2G and F3OPD regions, numeric}
  \item{F2D.F3OPD}{Functional connectivity between F2D and F3OPD regions, numeric}
  \item{F2OG.F3OPD}{Functional connectivity between F2OG and F3OPD regions, numeric}
  \item{F2OD.F3OPD}{Functional connectivity between F2OD and F3OPD regions, numeric}
  \item{F3OPG.F3OPD}{Functional connectivity between F3OPG and F3OPD regions, numeric}
  \item{FAG.F3TG}{Functional connectivity between FAG and F3TG regions, numeric}
  \item{FAD.F3TG}{Functional connectivity between FAD and F3TG regions, numeric}
  \item{F1G.F3TG}{Functional connectivity between F1G and F3TG regions, numeric}
  \item{F1D.F3TG}{Functional connectivity between F1D and F3TG regions, numeric}
  \item{F1OG.F3TG}{Functional connectivity between F1OG and F3TG regions, numeric}
  \item{F1OD.F3TG}{Functional connectivity between F1OD and F3TG regions, numeric}
  \item{F2G.F3TG}{Functional connectivity between F2G and F3TG regions, numeric}
  \item{F2D.F3TG}{Functional connectivity between F2D and F3TG regions, numeric}
  \item{F2OG.F3TG}{Functional connectivity between F2OG and F3TG regions, numeric}
  \item{F2OD.F3TG}{Functional connectivity between F2OD and F3TG regions, numeric}
  \item{F3OPG.F3TG}{Functional connectivity between F3OPG and F3TG regions, numeric}
  \item{F3OPD.F3TG}{Functional connectivity between F3OPD and F3TG regions, numeric}
  \item{FAG.F3TD}{Functional connectivity between FAG and F3TD regions, numeric}
  \item{FAD.F3TD}{Functional connectivity between FAD and F3TD regions, numeric}
  \item{F1G.F3TD}{Functional connectivity between F1G and F3TD regions, numeric}
  \item{F1D.F3TD}{Functional connectivity between F1D and F3TD regions, numeric}
  \item{F1OG.F3TD}{Functional connectivity between F1OG and F3TD regions, numeric}
  \item{F1OD.F3TD}{Functional connectivity between F1OD and F3TD regions, numeric}
  \item{F2G.F3TD}{Functional connectivity between F2G and F3TD regions, numeric}
  \item{F2D.F3TD}{Functional connectivity between F2D and F3TD regions, numeric}
  \item{F2OG.F3TD}{Functional connectivity between F2OG and F3TD regions, numeric}
  \item{F2OD.F3TD}{Functional connectivity between F2OD and F3TD regions, numeric}
  \item{F3OPG.F3TD}{Functional connectivity between F3OPG and F3TD regions, numeric}
  \item{F3OPD.F3TD}{Functional connectivity between F3OPD and F3TD regions, numeric}
  \item{F3TG.F3TD}{Functional connectivity between F3TG and F3TD regions, numeric}
  \item{FAG.F3OG}{Functional connectivity between FAG and F3OG regions, numeric}
  \item{FAD.F3OG}{Functional connectivity between FAD and F3OG regions, numeric}
  \item{F1G.F3OG}{Functional connectivity between F1G and F3OG regions, numeric}
  \item{F1D.F3OG}{Functional connectivity between F1D and F3OG regions, numeric}
  \item{F1OG.F3OG}{Functional connectivity between F1OG and F3OG regions, numeric}
  \item{F1OD.F3OG}{Functional connectivity between F1OD and F3OG regions, numeric}
  \item{F2G.F3OG}{Functional connectivity between F2G and F3OG regions, numeric}
  \item{F2D.F3OG}{Functional connectivity between F2D and F3OG regions, numeric}
  \item{F2OG.F3OG}{Functional connectivity between F2OG and F3OG regions, numeric}
  \item{F2OD.F3OG}{Functional connectivity between F2OD and F3OG regions, numeric}
  \item{F3OPG.F3OG}{Functional connectivity between F3OPG and F3OG regions, numeric}
  \item{F3OPD.F3OG}{Functional connectivity between F3OPD and F3OG regions, numeric}
  \item{F3TG.F3OG}{Functional connectivity between F3TG and F3OG regions, numeric}
  \item{F3TD.F3OG}{Functional connectivity between F3TD and F3OG regions, numeric}
  \item{FAG.F3OD}{Functional connectivity between FAG and F3OD regions, numeric}
  \item{FAD.F3OD}{Functional connectivity between FAD and F3OD regions, numeric}
  \item{F1G.F3OD}{Functional connectivity between F1G and F3OD regions, numeric}
  \item{F1D.F3OD}{Functional connectivity between F1D and F3OD regions, numeric}
  \item{F1OG.F3OD}{Functional connectivity between F1OG and F3OD regions, numeric}
  \item{F1OD.F3OD}{Functional connectivity between F1OD and F3OD regions, numeric}
  \item{F2G.F3OD}{Functional connectivity between F2G and F3OD regions, numeric}
  \item{F2D.F3OD}{Functional connectivity between F2D and F3OD regions, numeric}
  \item{F2OG.F3OD}{Functional connectivity between F2OG and F3OD regions, numeric}
  \item{F2OD.F3OD}{Functional connectivity between F2OD and F3OD regions, numeric}
  \item{F3OPG.F3OD}{Functional connectivity between F3OPG and F3OD regions, numeric}
  \item{F3OPD.F3OD}{Functional connectivity between F3OPD and F3OD regions, numeric}
  \item{F3TG.F3OD}{Functional connectivity between F3TG and F3OD regions, numeric}
  \item{F3TD.F3OD}{Functional connectivity between F3TD and F3OD regions, numeric}
  \item{F3OG.F3OD}{Functional connectivity between F3OG and F3OD regions, numeric}
  \item{FAG.ORG}{Functional connectivity between FAG and ORG regions, numeric}
  \item{FAD.ORG}{Functional connectivity between FAD and ORG regions, numeric}
  \item{F1G.ORG}{Functional connectivity between F1G and ORG regions, numeric}
  \item{F1D.ORG}{Functional connectivity between F1D and ORG regions, numeric}
  \item{F1OG.ORG}{Functional connectivity between F1OG and ORG regions, numeric}
  \item{F1OD.ORG}{Functional connectivity between F1OD and ORG regions, numeric}
  \item{F2G.ORG}{Functional connectivity between F2G and ORG regions, numeric}
  \item{F2D.ORG}{Functional connectivity between F2D and ORG regions, numeric}
  \item{F2OG.ORG}{Functional connectivity between F2OG and ORG regions, numeric}
  \item{F2OD.ORG}{Functional connectivity between F2OD and ORG regions, numeric}
  \item{F3OPG.ORG}{Functional connectivity between F3OPG and ORG regions, numeric}
  \item{F3OPD.ORG}{Functional connectivity between F3OPD and ORG regions, numeric}
  \item{F3TG.ORG}{Functional connectivity between F3TG and ORG regions, numeric}
  \item{F3TD.ORG}{Functional connectivity between F3TD and ORG regions, numeric}
  \item{F3OG.ORG}{Functional connectivity between F3OG and ORG regions, numeric}
  \item{F3OD.ORG}{Functional connectivity between F3OD and ORG regions, numeric}
  \item{FAG.ORD}{Functional connectivity between FAG and ORD regions, numeric}
  \item{FAD.ORD}{Functional connectivity between FAD and ORD regions, numeric}
  \item{F1G.ORD}{Functional connectivity between F1G and ORD regions, numeric}
  \item{F1D.ORD}{Functional connectivity between F1D and ORD regions, numeric}
  \item{F1OG.ORD}{Functional connectivity between F1OG and ORD regions, numeric}
  \item{F1OD.ORD}{Functional connectivity between F1OD and ORD regions, numeric}
  \item{F2G.ORD}{Functional connectivity between F2G and ORD regions, numeric}
  \item{F2D.ORD}{Functional connectivity between F2D and ORD regions, numeric}
  \item{F2OG.ORD}{Functional connectivity between F2OG and ORD regions, numeric}
  \item{F2OD.ORD}{Functional connectivity between F2OD and ORD regions, numeric}
  \item{F3OPG.ORD}{Functional connectivity between F3OPG and ORD regions, numeric}
  \item{F3OPD.ORD}{Functional connectivity between F3OPD and ORD regions, numeric}
  \item{F3TG.ORD}{Functional connectivity between F3TG and ORD regions, numeric}
  \item{F3TD.ORD}{Functional connectivity between F3TD and ORD regions, numeric}
  \item{F3OG.ORD}{Functional connectivity between F3OG and ORD regions, numeric}
  \item{F3OD.ORD}{Functional connectivity between F3OD and ORD regions, numeric}
  \item{ORG.ORD}{Functional connectivity between ORG and ORD regions, numeric}
  \item{FAG.SMAG}{Functional connectivity between FAG and SMAG regions, numeric}
  \item{FAD.SMAG}{Functional connectivity between FAD and SMAG regions, numeric}
  \item{F1G.SMAG}{Functional connectivity between F1G and SMAG regions, numeric}
  \item{F1D.SMAG}{Functional connectivity between F1D and SMAG regions, numeric}
  \item{F1OG.SMAG}{Functional connectivity between F1OG and SMAG regions, numeric}
  \item{F1OD.SMAG}{Functional connectivity between F1OD and SMAG regions, numeric}
  \item{F2G.SMAG}{Functional connectivity between F2G and SMAG regions, numeric}
  \item{F2D.SMAG}{Functional connectivity between F2D and SMAG regions, numeric}
  \item{F2OG.SMAG}{Functional connectivity between F2OG and SMAG regions, numeric}
  \item{F2OD.SMAG}{Functional connectivity between F2OD and SMAG regions, numeric}
  \item{F3OPG.SMAG}{Functional connectivity between F3OPG and SMAG regions, numeric}
  \item{F3OPD.SMAG}{Functional connectivity between F3OPD and SMAG regions, numeric}
  \item{F3TG.SMAG}{Functional connectivity between F3TG and SMAG regions, numeric}
  \item{F3TD.SMAG}{Functional connectivity between F3TD and SMAG regions, numeric}
  \item{F3OG.SMAG}{Functional connectivity between F3OG and SMAG regions, numeric}
  \item{F3OD.SMAG}{Functional connectivity between F3OD and SMAG regions, numeric}
  \item{ORG.SMAG}{Functional connectivity between ORG and SMAG regions, numeric}
  \item{ORD.SMAG}{Functional connectivity between ORD and SMAG regions, numeric}
  \item{FAG.SMAD}{Functional connectivity between FAG and SMAD regions, numeric}
  \item{FAD.SMAD}{Functional connectivity between FAD and SMAD regions, numeric}
  \item{F1G.SMAD}{Functional connectivity between F1G and SMAD regions, numeric}
  \item{F1D.SMAD}{Functional connectivity between F1D and SMAD regions, numeric}
  \item{F1OG.SMAD}{Functional connectivity between F1OG and SMAD regions, numeric}
  \item{F1OD.SMAD}{Functional connectivity between F1OD and SMAD regions, numeric}
  \item{F2G.SMAD}{Functional connectivity between F2G and SMAD regions, numeric}
  \item{F2D.SMAD}{Functional connectivity between F2D and SMAD regions, numeric}
  \item{F2OG.SMAD}{Functional connectivity between F2OG and SMAD regions, numeric}
  \item{F2OD.SMAD}{Functional connectivity between F2OD and SMAD regions, numeric}
  \item{F3OPG.SMAD}{Functional connectivity between F3OPG and SMAD regions, numeric}
  \item{F3OPD.SMAD}{Functional connectivity between F3OPD and SMAD regions, numeric}
  \item{F3TG.SMAD}{Functional connectivity between F3TG and SMAD regions, numeric}
  \item{F3TD.SMAD}{Functional connectivity between F3TD and SMAD regions, numeric}
  \item{F3OG.SMAD}{Functional connectivity between F3OG and SMAD regions, numeric}
  \item{F3OD.SMAD}{Functional connectivity between F3OD and SMAD regions, numeric}
  \item{ORG.SMAD}{Functional connectivity between ORG and SMAD regions, numeric}
  \item{ORD.SMAD}{Functional connectivity between ORD and SMAD regions, numeric}
  \item{SMAG.SMAD}{Functional connectivity between SMAG and SMAD regions, numeric}
  \item{FAG.COBG}{Functional connectivity between FAG and COBG regions, numeric}
  \item{FAD.COBG}{Functional connectivity between FAD and COBG regions, numeric}
  \item{F1G.COBG}{Functional connectivity between F1G and COBG regions, numeric}
  \item{F1D.COBG}{Functional connectivity between F1D and COBG regions, numeric}
  \item{F1OG.COBG}{Functional connectivity between F1OG and COBG regions, numeric}
  \item{F1OD.COBG}{Functional connectivity between F1OD and COBG regions, numeric}
  \item{F2G.COBG}{Functional connectivity between F2G and COBG regions, numeric}
  \item{F2D.COBG}{Functional connectivity between F2D and COBG regions, numeric}
  \item{F2OG.COBG}{Functional connectivity between F2OG and COBG regions, numeric}
  \item{F2OD.COBG}{Functional connectivity between F2OD and COBG regions, numeric}
  \item{F3OPG.COBG}{Functional connectivity between F3OPG and COBG regions, numeric}
  \item{F3OPD.COBG}{Functional connectivity between F3OPD and COBG regions, numeric}
  \item{F3TG.COBG}{Functional connectivity between F3TG and COBG regions, numeric}
  \item{F3TD.COBG}{Functional connectivity between F3TD and COBG regions, numeric}
  \item{F3OG.COBG}{Functional connectivity between F3OG and COBG regions, numeric}
  \item{F3OD.COBG}{Functional connectivity between F3OD and COBG regions, numeric}
  \item{ORG.COBG}{Functional connectivity between ORG and COBG regions, numeric}
  \item{ORD.COBG}{Functional connectivity between ORD and COBG regions, numeric}
  \item{SMAG.COBG}{Functional connectivity between SMAG and COBG regions, numeric}
  \item{SMAD.COBG}{Functional connectivity between SMAD and COBG regions, numeric}
  \item{FAG.COBD}{Functional connectivity between FAG and COBD regions, numeric}
  \item{FAD.COBD}{Functional connectivity between FAD and COBD regions, numeric}
  \item{F1G.COBD}{Functional connectivity between F1G and COBD regions, numeric}
  \item{F1D.COBD}{Functional connectivity between F1D and COBD regions, numeric}
  \item{F1OG.COBD}{Functional connectivity between F1OG and COBD regions, numeric}
  \item{F1OD.COBD}{Functional connectivity between F1OD and COBD regions, numeric}
  \item{F2G.COBD}{Functional connectivity between F2G and COBD regions, numeric}
  \item{F2D.COBD}{Functional connectivity between F2D and COBD regions, numeric}
  \item{F2OG.COBD}{Functional connectivity between F2OG and COBD regions, numeric}
  \item{F2OD.COBD}{Functional connectivity between F2OD and COBD regions, numeric}
  \item{F3OPG.COBD}{Functional connectivity between F3OPG and COBD regions, numeric}
  \item{F3OPD.COBD}{Functional connectivity between F3OPD and COBD regions, numeric}
  \item{F3TG.COBD}{Functional connectivity between F3TG and COBD regions, numeric}
  \item{F3TD.COBD}{Functional connectivity between F3TD and COBD regions, numeric}
  \item{F3OG.COBD}{Functional connectivity between F3OG and COBD regions, numeric}
  \item{F3OD.COBD}{Functional connectivity between F3OD and COBD regions, numeric}
  \item{ORG.COBD}{Functional connectivity between ORG and COBD regions, numeric}
  \item{ORD.COBD}{Functional connectivity between ORD and COBD regions, numeric}
  \item{SMAG.COBD}{Functional connectivity between SMAG and COBD regions, numeric}
  \item{SMAD.COBD}{Functional connectivity between SMAD and COBD regions, numeric}
  \item{COBG.COBD}{Functional connectivity between COBG and COBD regions, numeric}
  \item{FAG.FMG}{Functional connectivity between FAG and FMG regions, numeric}
  \item{FAD.FMG}{Functional connectivity between FAD and FMG regions, numeric}
  \item{F1G.FMG}{Functional connectivity between F1G and FMG regions, numeric}
  \item{F1D.FMG}{Functional connectivity between F1D and FMG regions, numeric}
  \item{F1OG.FMG}{Functional connectivity between F1OG and FMG regions, numeric}
  \item{F1OD.FMG}{Functional connectivity between F1OD and FMG regions, numeric}
  \item{F2G.FMG}{Functional connectivity between F2G and FMG regions, numeric}
  \item{F2D.FMG}{Functional connectivity between F2D and FMG regions, numeric}
  \item{F2OG.FMG}{Functional connectivity between F2OG and FMG regions, numeric}
  \item{F2OD.FMG}{Functional connectivity between F2OD and FMG regions, numeric}
  \item{F3OPG.FMG}{Functional connectivity between F3OPG and FMG regions, numeric}
  \item{F3OPD.FMG}{Functional connectivity between F3OPD and FMG regions, numeric}
  \item{F3TG.FMG}{Functional connectivity between F3TG and FMG regions, numeric}
  \item{F3TD.FMG}{Functional connectivity between F3TD and FMG regions, numeric}
  \item{F3OG.FMG}{Functional connectivity between F3OG and FMG regions, numeric}
  \item{F3OD.FMG}{Functional connectivity between F3OD and FMG regions, numeric}
  \item{ORG.FMG}{Functional connectivity between ORG and FMG regions, numeric}
  \item{ORD.FMG}{Functional connectivity between ORD and FMG regions, numeric}
  \item{SMAG.FMG}{Functional connectivity between SMAG and FMG regions, numeric}
  \item{SMAD.FMG}{Functional connectivity between SMAD and FMG regions, numeric}
  \item{COBG.FMG}{Functional connectivity between COBG and FMG regions, numeric}
  \item{COBD.FMG}{Functional connectivity between COBD and FMG regions, numeric}
  \item{FAG.FMD}{Functional connectivity between FAG and FMD regions, numeric}
  \item{FAD.FMD}{Functional connectivity between FAD and FMD regions, numeric}
  \item{F1G.FMD}{Functional connectivity between F1G and FMD regions, numeric}
  \item{F1D.FMD}{Functional connectivity between F1D and FMD regions, numeric}
  \item{F1OG.FMD}{Functional connectivity between F1OG and FMD regions, numeric}
  \item{F1OD.FMD}{Functional connectivity between F1OD and FMD regions, numeric}
  \item{F2G.FMD}{Functional connectivity between F2G and FMD regions, numeric}
  \item{F2D.FMD}{Functional connectivity between F2D and FMD regions, numeric}
  \item{F2OG.FMD}{Functional connectivity between F2OG and FMD regions, numeric}
  \item{F2OD.FMD}{Functional connectivity between F2OD and FMD regions, numeric}
  \item{F3OPG.FMD}{Functional connectivity between F3OPG and FMD regions, numeric}
  \item{F3OPD.FMD}{Functional connectivity between F3OPD and FMD regions, numeric}
  \item{F3TG.FMD}{Functional connectivity between F3TG and FMD regions, numeric}
  \item{F3TD.FMD}{Functional connectivity between F3TD and FMD regions, numeric}
  \item{F3OG.FMD}{Functional connectivity between F3OG and FMD regions, numeric}
  \item{F3OD.FMD}{Functional connectivity between F3OD and FMD regions, numeric}
  \item{ORG.FMD}{Functional connectivity between ORG and FMD regions, numeric}
  \item{ORD.FMD}{Functional connectivity between ORD and FMD regions, numeric}
  \item{SMAG.FMD}{Functional connectivity between SMAG and FMD regions, numeric}
  \item{SMAD.FMD}{Functional connectivity between SMAD and FMD regions, numeric}
  \item{COBG.FMD}{Functional connectivity between COBG and FMD regions, numeric}
  \item{COBD.FMD}{Functional connectivity between COBD and FMD regions, numeric}
  \item{FMG.FMD}{Functional connectivity between FMG and FMD regions, numeric}
  \item{FAG.FMOG}{Functional connectivity between FAG and FMOG regions, numeric}
  \item{FAD.FMOG}{Functional connectivity between FAD and FMOG regions, numeric}
  \item{F1G.FMOG}{Functional connectivity between F1G and FMOG regions, numeric}
  \item{F1D.FMOG}{Functional connectivity between F1D and FMOG regions, numeric}
  \item{F1OG.FMOG}{Functional connectivity between F1OG and FMOG regions, numeric}
  \item{F1OD.FMOG}{Functional connectivity between F1OD and FMOG regions, numeric}
  \item{F2G.FMOG}{Functional connectivity between F2G and FMOG regions, numeric}
  \item{F2D.FMOG}{Functional connectivity between F2D and FMOG regions, numeric}
  \item{F2OG.FMOG}{Functional connectivity between F2OG and FMOG regions, numeric}
  \item{F2OD.FMOG}{Functional connectivity between F2OD and FMOG regions, numeric}
  \item{F3OPG.FMOG}{Functional connectivity between F3OPG and FMOG regions, numeric}
  \item{F3OPD.FMOG}{Functional connectivity between F3OPD and FMOG regions, numeric}
  \item{F3TG.FMOG}{Functional connectivity between F3TG and FMOG regions, numeric}
  \item{F3TD.FMOG}{Functional connectivity between F3TD and FMOG regions, numeric}
  \item{F3OG.FMOG}{Functional connectivity between F3OG and FMOG regions, numeric}
  \item{F3OD.FMOG}{Functional connectivity between F3OD and FMOG regions, numeric}
  \item{ORG.FMOG}{Functional connectivity between ORG and FMOG regions, numeric}
  \item{ORD.FMOG}{Functional connectivity between ORD and FMOG regions, numeric}
  \item{SMAG.FMOG}{Functional connectivity between SMAG and FMOG regions, numeric}
  \item{SMAD.FMOG}{Functional connectivity between SMAD and FMOG regions, numeric}
  \item{COBG.FMOG}{Functional connectivity between COBG and FMOG regions, numeric}
  \item{COBD.FMOG}{Functional connectivity between COBD and FMOG regions, numeric}
  \item{FMG.FMOG}{Functional connectivity between FMG and FMOG regions, numeric}
  \item{FMD.FMOG}{Functional connectivity between FMD and FMOG regions, numeric}
  \item{FAG.FMOD}{Functional connectivity between FAG and FMOD regions, numeric}
  \item{FAD.FMOD}{Functional connectivity between FAD and FMOD regions, numeric}
  \item{F1G.FMOD}{Functional connectivity between F1G and FMOD regions, numeric}
  \item{F1D.FMOD}{Functional connectivity between F1D and FMOD regions, numeric}
  \item{F1OG.FMOD}{Functional connectivity between F1OG and FMOD regions, numeric}
  \item{F1OD.FMOD}{Functional connectivity between F1OD and FMOD regions, numeric}
  \item{F2G.FMOD}{Functional connectivity between F2G and FMOD regions, numeric}
  \item{F2D.FMOD}{Functional connectivity between F2D and FMOD regions, numeric}
  \item{F2OG.FMOD}{Functional connectivity between F2OG and FMOD regions, numeric}
  \item{F2OD.FMOD}{Functional connectivity between F2OD and FMOD regions, numeric}
  \item{F3OPG.FMOD}{Functional connectivity between F3OPG and FMOD regions, numeric}
  \item{F3OPD.FMOD}{Functional connectivity between F3OPD and FMOD regions, numeric}
  \item{F3TG.FMOD}{Functional connectivity between F3TG and FMOD regions, numeric}
  \item{F3TD.FMOD}{Functional connectivity between F3TD and FMOD regions, numeric}
  \item{F3OG.FMOD}{Functional connectivity between F3OG and FMOD regions, numeric}
  \item{F3OD.FMOD}{Functional connectivity between F3OD and FMOD regions, numeric}
  \item{ORG.FMOD}{Functional connectivity between ORG and FMOD regions, numeric}
  \item{ORD.FMOD}{Functional connectivity between ORD and FMOD regions, numeric}
  \item{SMAG.FMOD}{Functional connectivity between SMAG and FMOD regions, numeric}
  \item{SMAD.FMOD}{Functional connectivity between SMAD and FMOD regions, numeric}
  \item{COBG.FMOD}{Functional connectivity between COBG and FMOD regions, numeric}
  \item{COBD.FMOD}{Functional connectivity between COBD and FMOD regions, numeric}
  \item{FMG.FMOD}{Functional connectivity between FMG and FMOD regions, numeric}
  \item{FMD.FMOD}{Functional connectivity between FMD and FMOD regions, numeric}
  \item{FMOG.FMOD}{Functional connectivity between FMOG and FMOD regions, numeric}
  \item{FAG.GRG}{Functional connectivity between FAG and GRG regions, numeric}
  \item{FAD.GRG}{Functional connectivity between FAD and GRG regions, numeric}
  \item{F1G.GRG}{Functional connectivity between F1G and GRG regions, numeric}
  \item{F1D.GRG}{Functional connectivity between F1D and GRG regions, numeric}
  \item{F1OG.GRG}{Functional connectivity between F1OG and GRG regions, numeric}
  \item{F1OD.GRG}{Functional connectivity between F1OD and GRG regions, numeric}
  \item{F2G.GRG}{Functional connectivity between F2G and GRG regions, numeric}
  \item{F2D.GRG}{Functional connectivity between F2D and GRG regions, numeric}
  \item{F2OG.GRG}{Functional connectivity between F2OG and GRG regions, numeric}
  \item{F2OD.GRG}{Functional connectivity between F2OD and GRG regions, numeric}
  \item{F3OPG.GRG}{Functional connectivity between F3OPG and GRG regions, numeric}
  \item{F3OPD.GRG}{Functional connectivity between F3OPD and GRG regions, numeric}
  \item{F3TG.GRG}{Functional connectivity between F3TG and GRG regions, numeric}
  \item{F3TD.GRG}{Functional connectivity between F3TD and GRG regions, numeric}
  \item{F3OG.GRG}{Functional connectivity between F3OG and GRG regions, numeric}
  \item{F3OD.GRG}{Functional connectivity between F3OD and GRG regions, numeric}
  \item{ORG.GRG}{Functional connectivity between ORG and GRG regions, numeric}
  \item{ORD.GRG}{Functional connectivity between ORD and GRG regions, numeric}
  \item{SMAG.GRG}{Functional connectivity between SMAG and GRG regions, numeric}
  \item{SMAD.GRG}{Functional connectivity between SMAD and GRG regions, numeric}
  \item{COBG.GRG}{Functional connectivity between COBG and GRG regions, numeric}
  \item{COBD.GRG}{Functional connectivity between COBD and GRG regions, numeric}
  \item{FMG.GRG}{Functional connectivity between FMG and GRG regions, numeric}
  \item{FMD.GRG}{Functional connectivity between FMD and GRG regions, numeric}
  \item{FMOG.GRG}{Functional connectivity between FMOG and GRG regions, numeric}
  \item{FMOD.GRG}{Functional connectivity between FMOD and GRG regions, numeric}
  \item{FAG.GRD}{Functional connectivity between FAG and GRD regions, numeric}
  \item{FAD.GRD}{Functional connectivity between FAD and GRD regions, numeric}
  \item{F1G.GRD}{Functional connectivity between F1G and GRD regions, numeric}
  \item{F1D.GRD}{Functional connectivity between F1D and GRD regions, numeric}
  \item{F1OG.GRD}{Functional connectivity between F1OG and GRD regions, numeric}
  \item{F1OD.GRD}{Functional connectivity between F1OD and GRD regions, numeric}
  \item{F2G.GRD}{Functional connectivity between F2G and GRD regions, numeric}
  \item{F2D.GRD}{Functional connectivity between F2D and GRD regions, numeric}
  \item{F2OG.GRD}{Functional connectivity between F2OG and GRD regions, numeric}
  \item{F2OD.GRD}{Functional connectivity between F2OD and GRD regions, numeric}
  \item{F3OPG.GRD}{Functional connectivity between F3OPG and GRD regions, numeric}
  \item{F3OPD.GRD}{Functional connectivity between F3OPD and GRD regions, numeric}
  \item{F3TG.GRD}{Functional connectivity between F3TG and GRD regions, numeric}
  \item{F3TD.GRD}{Functional connectivity between F3TD and GRD regions, numeric}
  \item{F3OG.GRD}{Functional connectivity between F3OG and GRD regions, numeric}
  \item{F3OD.GRD}{Functional connectivity between F3OD and GRD regions, numeric}
  \item{ORG.GRD}{Functional connectivity between ORG and GRD regions, numeric}
  \item{ORD.GRD}{Functional connectivity between ORD and GRD regions, numeric}
  \item{SMAG.GRD}{Functional connectivity between SMAG and GRD regions, numeric}
  \item{SMAD.GRD}{Functional connectivity between SMAD and GRD regions, numeric}
  \item{COBG.GRD}{Functional connectivity between COBG and GRD regions, numeric}
  \item{COBD.GRD}{Functional connectivity between COBD and GRD regions, numeric}
  \item{FMG.GRD}{Functional connectivity between FMG and GRD regions, numeric}
  \item{FMD.GRD}{Functional connectivity between FMD and GRD regions, numeric}
  \item{FMOG.GRD}{Functional connectivity between FMOG and GRD regions, numeric}
  \item{FMOD.GRD}{Functional connectivity between FMOD and GRD regions, numeric}
  \item{GRG.GRD}{Functional connectivity between GRG and GRD regions, numeric}
}}
\source{
\url{https://fcon_1000.projects.nitrc.org/indi/adhd200/}
}
\usage{
frontal2D
}
\description{
A dataset containing the functional connectivity between
frontal lobe areas of the brain in 24 control and 24 patients with
Attention-Deficit/Hyperactivity Disorder (ADHD).
}
\details{
Data was taken from the ADHD200 dataset and variables were
manipulated in order to be different of the original data.
Functional connectivity was measure as the Pearson correlation
between the average fMRI signal from the regions of interest (ROI)
within 28 anatomical areas of the frontal lobe. Then, a total of
378 pairwise connections are contained in the dataset.
}
\keyword{datasets}
