% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{relatorio}
\alias{relatorio}
\title{Relatório NBR}
\usage{
relatorio(link, root = "Rel_NBR-Termico")
}
\arguments{
\item{link}{"The google sheet sharing link"}

\item{root}{"The name of the root folder to store all output}
}
\value{
Several plots and tables as a result of the NBR norm.
}
\description{
Relatório NBR
}
